requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
playwright>=1.32.0
aiohttp>=3.8.4
python-dotenv>=1.0.0
pydantic>=1.10.7
httpx>=0.24.0